Convert to pure ES6 project
react-ol now exports as an ES6 only project, with peer dependencies of
ol, react, and react-dom. This way, when the importing package compiles
its ES6, it will also compile this code with it, and thus use the same
ol module.

We can't include any babel tools to make it both ES6 and ES5 because on
import, the parent package's babel will go to this package's babel
tools, and get all confused. Therefore pure ES6 it is.

Curse you, openlayers, for your ES6 plague.